se cambia el tipo de role_id de tinyint a int
El id de la tabla roles se genera como int, así que la columna role_id
de users ahora usa el mismo tipo y la clave foránea queda consistente.

+++ src/users/entities/user.entity.ts
@@ -31,7 +31,7 @@ export class User {
   @Column({ type: 'varchar', name: 'is_active', default: true })
   isActive: boolean | undefined;
 
-  @Column({ type: 'tinyint', name: 'role_id' })
+  @Column({ type: 'int', name: 'role_id' })
   roleId: number | undefined;
 
   @CreateDateColumn({ name: 'created_at' })
